@@ -13,21 +13,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^19.0.0",
+    "@angular/animations": "^20.1.6",
     "@angular/cdk": "^19.2.19",
-    "@angular/common": "^19.0.0",
-    "@angular/compiler": "^19.0.0",
-    "@angular/core": "^19.0.0",
-    "@angular/forms": "^19.0.0",
+    "@angular/common": "^20.1.6",
+    "@angular/compiler": "^20.1.6",
+    "@angular/core": "^20.1.6",
+    "@angular/fire": "^20.0.1",
+    "@angular/forms": "^20.1.6",
     "@angular/material": "^19.2.19",
-    "@angular/platform-browser": "^19.0.0",
-    "@angular/platform-browser-dynamic": "^19.0.0",
-    "@angular/router": "^19.0.0",
+    "@angular/platform-browser": "^20.1.6",
+    "@angular/platform-browser-dynamic": "^20.1.6",
+    "@angular/router": "^20.1.6",
+    "@capacitor-community/firebase-analytics": "^7.0.1",
+    "@firebasegen/default-connector": "file:dataconnect-generated/js/default-connector",
     "@fortawesome/fontawesome-free": "^6.7.2",
     "@ionic/angular": "^8.0.0",
     "bootstrap": "^5.3.6",
     "fire": "^0.55.1",
-    "firebase": "^12.0.0",
+    "firebase": "^12.1.0",
     "ionicons": "^7.0.0",
     "rxjs": "~7.8.0",
     "swiper": "^11.2.8",
@@ -36,15 +39,15 @@
     "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.0.0",
     "@angular-eslint/builder": "^19.0.0",
     "@angular-eslint/eslint-plugin": "^19.0.0",
     "@angular-eslint/eslint-plugin-template": "^19.0.0",
     "@angular-eslint/schematics": "^19.0.0",
     "@angular-eslint/template-parser": "^19.0.0",
-    "@angular/cli": "^19.0.0",
-    "@angular/compiler-cli": "^19.0.0",
-    "@angular/language-service": "^19.0.0",
+    "@angular/build": "^20.1.5",
+    "@angular/cli": "^20.1.5",
+    "@angular/compiler-cli": "^20.1.6",
+    "@angular/language-service": "^20.1.6",
     "@ionic/angular-toolkit": "^12.0.0",
     "@types/google.maps": "^3.58.1",
     "@types/jasmine": "~5.1.0",
@@ -61,7 +64,7 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "typescript": "~5.6.3"
+    "typescript": "~5.8.3"
   },
   "description": "An Ionic project"
 }
