@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { NewsService } from '../news_services/news.service';
 import { Router } from '@angular/router';
+import { ActivityLoggerService } from '../App-services/activity-logger.service';
+import { LogService } from '../App-services/log.service';
 
 
 @Component({
@@ -13,13 +15,14 @@ export class LatestNewsPage implements OnInit {
  
 newsItems: any[] = [];
 
-  constructor(private newsService: NewsService, private router: Router) {}
-
+  constructor(private newsService: NewsService, private router: Router, private  logger:ActivityLoggerService, private logService: LogService) {}
   openNewsDetails(id: number) {
   this.router.navigate(['/news-details', id]);
+  //this.logger.logActivity('page_view', { page: 'latest-news' });
 }
 
   ngOnInit() {
+    this.logService.log('Page viewed','Latest News Page');
     this.newsService.getNews().subscribe((
       data: any[] )=> {
         console.log('News data:', data);
