@@ -1,3 +1,48 @@
-export const environment = {
-  production: true
+// import { initializeApp } from "@angular/fire/app";
+// import { getAnalytics } from "@angular/fire/analytics";
+
+// export const environment = {
+//   production: true,
+//   firebaseConfig: {
+//     apiKey: 'AIzaSyBklOBwfiSURtNgRlOHDQVC9xLHTxhsBq8',
+//     authDomain: 'nlsa-app.firebaseapp.com',
+//     projectId: 'nlsa-app',
+//     databaseURL: 'https://nlsa-app-default-rtdb.firebaseio.com',
+//     storageBucket: 'nlsa-app.firebasestorage.app',
+//     messagingSenderId: '692680829717',
+//     appId: '1:692680829717:web:88c9901f07bf990026171c',
+//     measurementId: 'G-2ZX5RGPVYN',
+//   }
+// };
+
+// const app = initializeApp(environment.firebaseConfig);
+
+// const analytics = getAnalytics(app);
+
+// Import the functions you need from the SDKs you need
+import { initializeApp } from "firebase/app";
+import { getAnalytics } from "firebase/analytics";
+// TODO: Add SDKs for Firebase products that you want to use
+// https://firebase.google.com/docs/web/setup#available-libraries
+
+// Your web app's Firebase configuration
+// For Firebase JS SDK v7.20.0 and later, measurementId is optional
+
+export const firebaseConfig = {
+  apiKey: "AIzaSyDwrevrwNl0BdbRE9-bBdzYG2X5ebXBXJI",
+  authDomain: "nlsa-firebaseapp.firebaseapp.com",
+  projectId: "nlsa-firebaseapp",
+  storageBucket: "nlsa-firebaseapp.firebasestorage.app",
+  messagingSenderId: "481002941013",
+  appId: "1:481002941013:web:bc626bd35d0aab6dce02dc",
+  measurementId: "G-F0NR4H802C"
 };
+
+export const environment = {
+  production: true,
+  firebaseConfig: firebaseConfig
+};
+
+// Initialize Firebase
+const app = initializeApp(firebaseConfig);
+const analytics = getAnalytics(app);
